Move method for writing files into file storage class
VDX-1934
